update for [W][W] and [F][F] Special Energy
Not yet implemented: Special Energy cards that can only be attached to specific Pokémon don't discard themselves when the Pokémon they're attached to is no longer valid. This is similar to Equip Cards in Yugioh that are destroyed due to game mechanics when their target is also no longer valid.